m10

+++ iCaRL_net.py
@@ -186,7 +186,7 @@ class iCaRL(nn.Module):
                 elif self.loss_config == 1:
                     #NLLL
                     out = torch.log(torch.softmax(out, dim=1))
-                    loss = self.clf_loss(out[:, self.n_known:self.n_classes], labels_hot[:, self.n_known:self.n_classes])
+                    loss = self.clf_loss(out[:, self.n_known:self.n_classes], labels_hot[:, self.n_known:self.n_classes].long())
 
                 elif self.loss_config == 2:
                     #L1
